Refactor mesh vector settings into generic VectorField settings classes
Split classes to separate files

Preserve API using deprecated subclasses with the old class names.
