+++ python/PyQt6/core/auto_additions/qgsvectorfieldtracessettings.py
@@ -1,0 +1,5 @@
+# The following has been generated automatically from src/core/vectorfield/qgsvectorfieldtracessettings.h
+try:
+    QgsVectorFieldTracesSettings.__group__ = ['vectorfield']
+except (NameError, AttributeError):
+    pass
